installer/requirements.txt
pysha3==1.0.2
coincurve==18.0.0